requirements.txt
pycryptodome==3.10.1
requests==2.25.1
